chore: split `OrderProgressBarV2` unit into smaller units
