+++ apps/cowswap-frontend/src/common/pure/OrderProgressBarV2/steps/ExecutingStep.tsx
@@ -1,0 +1,27 @@
+import { STEPS } from '../constants'
+import StepsWrapper from '../container/StepsWrapper'
+import * as styledEl from '../styled'
+
+interface ExecutingStepProps {
+  children: React.ReactNode
+}
+
+function ExecutingStep({ children }: ExecutingStepProps) {
+  return (
+    <styledEl.ProgressContainer>
+      {children}
+      <StepsWrapper
+        steps={STEPS}
+        currentStep={2}
+        customStepTitles={{ 2: 'Best price found!' }}
+        extraContent={
+          <styledEl.Description>
+            The winner of the competition is now executing your order on-chain.
+          </styledEl.Description>
+        }
+      />
+    </styledEl.ProgressContainer>
+  )
+}
+
+export default ExecutingStep
